requirements.txt
# her
dj-database-url==0.4.2
dj-static==0.0.6
django-bootstrap3==8.2.3
django-toolbelt==0.0.1
gunicorn==19.7.1

